requirements.txt
psutil==5.9.8
python-dotenv==1.0.1
rich==13.7.1
requests==2.31.0